Studio Library : Matrix Updates
The poses now save local and world space matrix information, and the pose application methods have been updated to use the stored matrices of the poses to better calculate the end-result value.  This is very messy and is only -just- working.  I just discovered I have broken the blending behavior and need to shuffle where the matrix behaviors are unpacked and passed to the setting functions.  So just checking in my work thus far-as a snapshot and while I uproot things to better work with how default studio library functions.