Matrix Work
The action of applying the pose and managing the matrices is now performed before the pose is handed off to the poser functions.  The current behavior is to apply each pose's matrix, capture the values via getattr, and restore the position.  With this, Maya does the work of figuring out the 100% transform values of the pose, and the pose applying is identical to how studio library already works.

The only hitch is that, it appears that the pose needs to be applied in its entirety, then the objects need to fetch their values, and finally the pose should then be restored.  The poser doesn't really work like this currently as each node pair handle their own posing. 

This checkin does implement a sorting block that poses the pairs starts at the top of the hierarchy, on down.  This didn't solve the issue with applying poses however, but I was hoping.  This will be removed on the next checking but this revision contains it, just in case it comes in handy in the future.