Matrix Transforms are working
Applying transforms from matrix data is now working correctly.   There is some clunky code where the pose is entirely applied to the rig, and the local transforms are queried off the nodes, and then the rig is restored to its previous state.  This could use a lot of optimizing and cleanup but in my testing, the functionality is wokring.