Matrix test method
Updating the hasTransforms() method to use the same self.isTransform(attr) check used elsewhere when fetching the pose data.  This will be used when loading the pose, for displaying to the user if the pose was saved with matrix information embedded.  If it wasn't, it is a legacy pose and the transforms will be posed like any other animatable channel.