Matrix Transforms Update
To support the use of matrices when posing transforms, the pose file contents were updated to contain both the local space matrix, and the world matrix of each object that included transforms in their attributes.  This update restores the behavior of saving the channel box values for transform attributes, but these values are not currently used when applying the pose.  This change is to cover any future bases where, perhaps, a "legacy" pose mode is added to the GUI, allowing the original behavior to be restored, instead of matrices being the permanent transform posing method.