Matrix Posing Initial Wrap-up
The posing using matrices has been mostly completed and bug tested.  The fallback behavior appears to be working well for older poses, and newer poses with matrices embedded are working as well.  There is definitely a bit of cleanup and optimization that should happen but this should get the animators up and running for now.