Disable auto-keyframe decorator added back
There isn't a great place to disable the auto-keyframe behavior as it will get turned on and off over and over, but I found a place in the code that doesn't get hit too often.  This will ensure that having Auto-Keyframe enabled in maya doesn't cause keyframe thrashing while applying and pose scrubbing.