MSN : Pose File Versioning
The internal version number for the pose objects have been exposed to the GUI, allowing users to see the version number of each pose.  A tooltip has been added to summarize the differences between pose versions.  The internal logic of actually applying poses is unchanged, and saving a pose now reads from a pose version variable instead of the original default of 1.0.0, but is otherwise unchanged.