Updating character pre-pose-application behavior
Prior to pose application, any IK controllers that may be in the pose, are removed from the pose object dictionary.  This is due to the rig being switched to FK prior to pose application, and with the limbs returning to their IK state, if they were enabled, after the pose finishes applying.  So posing the IK controllers is redundant, but also breaks the rig if using the "Apply Relative To..." behavior.