A logic bug was causing a warning to fire when it shouldn't.
The applyRelativeTo and pose version test weren't catching the situation where the user wasn't using Apply Relative To, while working with a new pose, causing the "this is an old pose" warning to show when the user wasn't even trying to use the relative pose application behavior.  Simply logic fix to prevent that unique combination of tic-boxes and pose file types from firing off a bogus warning.