Updating logic behind sending a rig to the "posing" state
The process of sending a rig to FK while posing, and back to IK, was encountering some fringe issues.  One example was applying a pose on some fingers, would cause the rig to flip the IKs to FK, and back again after the pose was applied.  This was unnecessary since nothing about the arms, or anything related, was being posed, so the switch to FK was laying down keys on all limbs when they weren't even being posed.

The pose logic will now bypass sending limbs to FK if Studio Library is not posing any part of the IK limb, or any of the limbs parents.