removing the before and after loads and replacing them with decorators
The beforeLoad and afterLoad functions performed two actions, restoring your selection, and wrapping everything inbetween with an undo chunk.  I have replaced both with decorators attached to the load() method, which helped clean up a bit of the code, and cleaned up a multi-undo-operation that needed to be performed to undo a pose application.