DG to Parallel Eval Mode Fix
For some reason, certain bones in the character's skeleton will frequently gimbal when a pose is applied, and parallel mode is re-enabled.  I found that hitting the transforms of these nodes with a set keyframe, after the posing has finished and right before switching back to DG mode, will alleviate this pop.