Legacy Rig Fix
Adding a clunky little function that I will rip out at the earliest possible moment, to fix an issue where Animators need some values from IK controllers, saved inside the Studio Library poses, but not others.  This is a temporary fix so I have tested the behavior and it is working well, but I will remove it as soon as possible to avoid fallout, or any potential triage that might be necessary to maintain it.