Adding a new behavior when posing IK systems of our rig.  In a prior update, Studio Library was changed to intentionally remove IK controllers or Pole Vectors, from the nodes being posed.  This was to prevent poses with IK controllers saved inside, from changing the ik states in another scene.  In tandem with that change, any IK system being posed is switched to FK mode, the pose is applied, and the IK is matched and re-enabled using the rig's only IK system.  If the limb was already in FK, Studio Library wouldn't do anything and would pose the selected parts of the limb as usual.
The issue came up where selecting only an IK controller and applying a pose, would result in Studio Library not doing anything.  This was because the IK control was getting removed from the pose list.  This obviously didn't make sense to the end users.

The fix for this scenario is to add the FK bones of any IK system the user is posing, to the nodes being posed.  Studio Library will pluck out the IK controller as usual, and flip the limb to FK, and continue on its way.